refactor(cli-core): CompilationUnit::fetch -> package_fetch::fetch_compilation_unit
The network-dep fetcher (HTTP + on-disk cache) used to live on
`CompilationUnit::fetch` behind `cfg(not(target_arch = "wasm32"))`.
Move it out as a free function `leo_cli_core::package_fetch::fetch_compilation_unit`.
Both direct callers (`upgrade.rs`, `common/query.rs`) and the
`fetch_network_dependency` path inside `Package::from_directory_impl`
now route through it.

To break the resulting circular-dep concern (crates/package needs to
fetch, but can't depend on leo-cli-core), `NetworkConfig` gains a
`fetcher: CompilationUnitFetcher` fn-pointer field. `Package::from_directory*`
and `from_directory_no_graph` take the fetcher as a new argument. CLI
callers pass `leo_cli_core::package_fetch::fetch_compilation_unit`;
the benchmarks crate passes the new `leo_package::reject_network_fetcher`
helper (local-only fixtures, no network deps expected).

Side effect: `find_cached_edition` and `parse_dependencies_from_aleo`
become `pub` in `leo-package` so the moved fetcher can call them.
Several remaining `cfg(not(target_arch = "wasm32"))` gates in
`crates/package` are dropped (Manifest::{read,write}_to_file,
Package::{from_directory_*}, CompilationUnit::{from_aleo_path,
from_package_path, from_test_path}, `home_path.canonicalize()` in
`from_directory_impl`) — those bodies use only `std::fs` and
`DiskFileSource`, both of which compile on wasm32.
