+++ crates/leo-cli-core/src/lib.rs
@@ -46,5 +46,6 @@ pub mod project;
 // of `crates/leo-package` so that crate stays purely wasm-buildable. Each
 // module is `#[cfg(not(target_arch = "wasm32"))]`-gated internally.
 pub mod network;
+pub mod package_fetch;
 pub mod package_init;
 pub mod validation;
