+++ crates/package/src/lib.rs
@@ -74,7 +74,6 @@
 pub mod errors;
 pub use errors::*;
 
-#[cfg(not(target_arch = "wasm32"))]
 use leo_ast::NetworkName;
 #[cfg(not(target_arch = "wasm32"))]
 use leo_errors::Backtraced;
